[iOS] Fix iOS and Windows builds
